add back the ReifyWindows to the GBK, some internal tests depend on it.
